+++ docs/src/pages/api/DateUtils.js
@@ -68,7 +68,8 @@ console.log(newRange.to) // 2015-05-24`}</CodeBlock>
         isDayAfter <code>(day1: Date, day2: Date) ⇒ boolean</code>
       </h3>
       <p>
-        Return <code>true</code> if <code>day1</code> is after <code>day2</code>.
+        Return <code>true</code> if <code>day1</code> is after <code>day2</code>
+        .
       </p>
 
       <h3>