+++ types/utils.d.ts
@@ -1,6 +1,6 @@
-// TypeScript Version: 2.2
+// TypeScript Version: 3.1
 
-import { RangeModifier, Modifier } from "./common";
+import { RangeModifier, Modifier } from './common';
 
 export interface LocaleUtils {
   formatDate(date: Date, format?: string | string[], locale?: string): string;
@@ -26,9 +26,9 @@ export interface LocaleUtils {
     string
   ];
   parseDate(str: string, format?: string, locale?: string): Date;
-}
+};
 
-export interface DateUtils {
+export const DateUtils: {
   addDayToRange(day: Date, range: RangeModifier): RangeModifier;
   addMonths(d: Date, n: number): Date;
   clone(d: Date): Date;
@@ -41,12 +41,12 @@ export interface DateUtils {
   isPastDay(day: Date): boolean;
   isSameDay(day1: Date, day2: Date): boolean;
   isSameMonth(day1: Date, day2: Date): boolean;
-}
+};
 
-export interface ModifiersUtils {
+export const ModifiersUtils: {
   dayMatchesModifier(day: Date, modifier?: Modifier | Modifier[]): boolean;
   getModifiersForDay(
     day: Date,
     modifiers: Record<string, Modifier | Modifier[]>
   ): string[];
-}
+};